Debian packaging: Require any of python[3/2]
Depends on any of the "python3", "python2" or "python" packages.

apm was recently updated to include newer node-gyp,
which supports both Python 2 and Python 3.
(Older version of node-gyp only supported Python 2).

node-gyp (and by extension, apm) looks for an executable named
"python", "python2", or "python3" on the PATH by default,
but the location can also be specified on the command-line
with "PYTHON=path/to/python" or "npm_config_python=path/to/python".