script: Download correct electron binaries
If users manually run `npm install`, we want to
make sure the correct Electron chromedriver and mksnapshot
are installed. The bootstrap script handles this usually,
but not for manual installs in the scripts dir.

This postinstall script conditionally runs, only if needed,
to correct downloading the wrong (v9.0.0/v9.0.2) binaries.