CI: don't worry about apm/package-lock.json
We can't expect apm/package-lock.json to be stable
when installing with `npm install --global-style`, apparently.

It's unstable, picking up any and every possible
(direct or indirect) dependency bump.

Delete that file from the cache key.

TODO: follow up with the npm maintainers
to see if this is considered a bug and can be fixed.