script/bootstrap: apm prints current Atom version
Set the ATOM_RESOURCE_PATH env variable in script/bootstrap,
so that `apm --version` can find the `package.json`
at the root of the repository, and print the version of Atom
that's currently being bootstrapped.

This used to print whatever Atom version
was installed to the user's system, which was a bit confusing
and not relevant at all to the bootstrapping process.