script/bootstrap: apm prints the current Atom ver
Lets apm know where this repository's Atom metadata is located,
so that `apm --version` prints the version of Atom being bootstrapped.

Previously, this would print "unknown" if no Atom was installed to the
system, or whatever stable version was installed...
which was irrelevant info during the bootstrapping process.

This should be more straightforward, and less confusing.