Sanitize "_" command-line flag, ignore non-strings
We expect these arguments to be paths or URLs (strings)
for Atom to open when it finishes initializing.

This commit ensures we do not do string-only operations
on these arguments if the arguments are not actually strings.

(Without it, we get TypeErrors and partial/incomplete startups.)

(Important for when macOS Gatekeeper adds a "-psn_0_[digits]" flag
to an app it has just verified or intercepted. (Why, Apple?))