spec: parseCommandLine() ignores macOS -psn_0 flag
This flag is set by Gatekeeper. Test to make sure we filter it out.

yargs thinks this is a bunch of short flags (-p -s -n -_ -0 -_ ...),
which means to add boolean "true"s to each supposed flag's array
of arguments, even adding boolean "true"s
to what is meant to be the"non-flag argument" array/ ("options._")

That gives our parser TypeErrors; It does string-only operations
on the elements from the "_" non-flag argument array.

If we filter this input properly, we won't have TypeErrors.