package-lock.json: Hoist tree-sitter-rust
Fixes CI.

(The package at packages/language-rust-bundled gets symlinked
into the main node_modules... It holds the actual tree-sitter-rust
package on disk in its own node_modules subdirectory.

This wasn't being restored when caching and restoring the main
node_modules folder. So tests would fail by not finding this package.
Hoist this so it is copied under the main node_modules folder,
not just symlinked in.)