CI: Use the Ubuntu 16.04 docker image for Linux CI
Allows us to keep building on Ubuntu 16.04, even after Azure Pipelines
removes the ability to use it as the host OS for CI. This is important
for using old enough `glibc` that will run on currently supported
Debian 9, but also allows stragglers still using Ubuntu 16.04 to run
the latest builds of Atom, for now.

We can use a newer Ubuntu host image (18.04, 20.04, etc.)
to run the older Ubuntu 16.04 OS as a Docker container.

(All of the job's steps will run in the Docker container,
rather than directly on the host.)

(Note: As an alternative to building on end-of-life Ubuntu 16.04,
we could instead build on the very slightly newer Debian 9 OS,
which is still supported; This would also be as a Docker container).